src/test/java/com/study/testcode/programmers/FillPuzzlePiecesNonCheatingTest.java
package com.study.testcode.programmers;

import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.TestPropertySource;

import static com.study.testcode.common.constant.DateConstants.CONFIG_TEST;
import static org.junit.jupiter.api.Assertions.assertArrayEquals;
import static org.junit.jupiter.api.Assertions.assertEquals;

@SpringBootTest
@TestPropertySource(locations = CONFIG_TEST)
public class FillPuzzlePiecesNonCheatingTest {
    @Autowired
    private FillPuzzlePiecesNonCheating fillPuzzlePiecesNonCheating;

    @Test
    void fillPuzzlePiecesNonCheatingTest() {
        int[][] game_board = {{1,1,0,0,1,0},{0,0,1,0,1,0},{0,1,1,0,0,1},{1,1,0,1,1,1},{1,0,0,0,1,0},{0,1,1,1,0,0}};
        int[][] table = {{1,0,0,1,1,0},{1,0,1,0,1,0},{0,1,1,0,1,1},{0,0,1,0,0,0},{1,1,0,1,1,0},{0,1,0,0,0,0}};
        int result = fillPuzzlePiecesNonCheating.solution(game_board, table);
        assertEquals(14, result);
    }
}
